// src/interfaces/Shop.ts
export interface Shop {
	createdAt: number
	id: number
	mexchangeStoreId: string
	name: string
	phone: string
	address: string
	storeOwnerName: string
	storeOwnerPhone: string
	storeOwnerId: string
	image: string
	// password?: string

	province: string
	district: string
	ward: string
	created: number
	updated_at: number
	partner_id: string
	partner_refer: string
	lat: number
	lng: number
	state: string
	username?: any
}